requirements.txt
nest_asyncio==1.5.1
Pillow==9.1.1
pyrlottie==2022.0.1
quart==0.17.0
SQLAlchemy==1.4.22
Telethon==1.24.0
